requirements.txt
streamlit
feedparser
groq
python-dotenv
PyPDF2
reportlab
pandas
psycopg2-binary
extra-streamlit-components